+++ MA4_files/person.py
@@ -10,7 +10,7 @@ class Person(object):
 		lib.Person_getAge.restype = ctypes.c_int
 		lib.Person_setAge.argtypes = [ctypes.c_void_p,ctypes.c_int]
 		lib.Person_getDecades.argtypes = [ctypes.c_void_p]
-		lib.Person_getDecades.restype = ctypes.c_int
+		lib.Person_getDecades.restype = ctypes.c_double
 		lib.Person_delete.argtypes = [ctypes.c_void_p]
 		self.obj = lib.Person_new(age)
